+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Modbus connection error which is raised if connection to the target device could not be established 
+- Unit tests for (un)successful modbus connection
+
 ### Changed
 - Logging levels for better usability of Logging functionality